pyproject.toml
[project]
name = "odin-bot"
version = "2.0.0"
description = "Odin — Discord moderation and utility bot with web dashboard"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "discord.py>=2.3,<3",
    "aiohttp>=3.9",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
    "ruff",
]

[project.scripts]
odin = "src.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
